# Conflicts:
#	casino-spring-mvc/src/main/java/com/company/controller/AccountController.java
#	casino-spring-mvc/src/main/java/com/company/models/account/User.java

+++ casino-spring-mvc/src/main/java/com/company/abstractions/IReadOnlyPlayingResult.java
@@ -1,8 +1,10 @@
 package com.company.abstractions;
 
+import java.math.BigDecimal;
+
 public interface IReadOnlyPlayingResult {
     public boolean isWin();
-    public double getWinPrice();
+    public BigDecimal getWinPrice();
     public long getBetId();
     public String getGameOutcomeView();
 }
