+++ casino-spring-mvc/src/main/java/com/company/models/view/EditOutcomeModel.java
@@ -4,10 +4,12 @@ import jakarta.validation.constraints.Min;
 import lombok.Data;
 import org.springframework.security.core.context.SecurityContextHolder;
 
+import java.math.BigDecimal;
+
 @Data
 public class EditOutcomeModel {
     private Long outcomeId;
     private String value;
     @Min(1)
-    private Double koef;
+    private BigDecimal koef;
 }
